better cmake for debug

+++ setup.py
@@ -4,21 +4,25 @@ from distutils.core import setup
 import subprocess
 import os
 
-def compile_C_code(verbose=0):
+def compile_C_code(verbose=0, build_type="Release"):
     """
     Compile automaticaly the C and Fortran code
     when installing the program with python setup.py
     """
+
+    if build_type not in ["Release", "Debug", "None"]:
+        raise ValueError("build type must be Release, Debug or None")
     ret = subprocess.call("rm -rf lib/build", shell = True)
 
     os.mkdir("lib/build")
     os.chdir("lib/build")
 
-    ret = subprocess.call("cmake ..", shell = True)
-    comp = "make VERBOSE={0}".format(verbose)
+    cmd = "cmake -DCMAKE_BUILD_TYPE=" + build_type + " .."
+    ret = subprocess.call(cmd, shell = True)
     if ret != 0:
         raise ValueError("cmake returned error {0}".format(ret))
-    ret = subprocess.call(comp, shell = True)
+    cmd = "make VERBOSE={0}".format(verbose)
+    ret = subprocess.call(cmd, shell = True)
     if ret != 0:
         raise ValueError("make returned error {0}".format(ret))
 
@@ -62,7 +66,7 @@ VERSION          = '1.3.0'
 
 compilation = True
 if compilation:
-    compile_C_code()
+    compile_C_code(build_type="Debug")
 
 setup(
     name=NAME,
